Adiciona ViewPager e corrije parse do GSON

+++ app/src/main/java/com/monit/coordinate/Coordinate.java
@@ -1,11 +1,11 @@
 package com.monit.coordinate;
 
-public class Coordinate {
+public class Coordinate{
 
-    public Long x;
+    public Double x;
     public Long y;
 
-    public Coordinate(Long x, Long y) {
+    public Coordinate(Double x, Long y) {
         this.x = x;
         this.y = y;
     }
